+++ internal/trace/parent/parent.go
@@ -17,19 +17,19 @@ package parent
 import (
 	"context"
 
-	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/label"
+	"go.opentelemetry.io/otel/trace"
 )
 
-func GetSpanContextAndLinks(ctx context.Context, ignoreContext bool) (otel.SpanContext, bool, []otel.Link) {
-	lsctx := otel.SpanContextFromContext(ctx)
-	rsctx := otel.RemoteSpanContextFromContext(ctx)
+func GetSpanContextAndLinks(ctx context.Context, ignoreContext bool) (trace.SpanContext, bool, []trace.Link) {
+	lsctx := trace.SpanContextFromContext(ctx)
+	rsctx := trace.RemoteSpanContextFromContext(ctx)
 
 	if ignoreContext {
 		links := addLinkIfValid(nil, lsctx, "current")
 		links = addLinkIfValid(links, rsctx, "remote")
 
-		return otel.SpanContext{}, false, links
+		return trace.SpanContext{}, false, links
 	}
 	if lsctx.IsValid() {
 		return lsctx, false, nil
@@ -37,14 +37,14 @@ func GetSpanContextAndLinks(ctx context.Context, ignoreContext bool) (otel.SpanC
 	if rsctx.IsValid() {
 		return rsctx, true, nil
 	}
-	return otel.SpanContext{}, false, nil
+	return trace.SpanContext{}, false, nil
 }
 
-func addLinkIfValid(links []otel.Link, sc otel.SpanContext, kind string) []otel.Link {
+func addLinkIfValid(links []trace.Link, sc trace.SpanContext, kind string) []trace.Link {
 	if !sc.IsValid() {
 		return links
 	}
-	return append(links, otel.Link{
+	return append(links, trace.Link{
 		SpanContext: sc,
 		Attributes: []label.KeyValue{
 			label.String("ignored-on-demand", kind),