+++ sdk/metric/aggregator/aggregatortest/test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package aggregatortest
+package aggregatortest // import "go.opentelemetry.io/otel/sdk/metric/aggregator/aggregatortest"
 
 import (
 	"context"